+++ library/common/src/main/java/com/google/android/exoplayer2/Player.java
@@ -1207,6 +1207,25 @@ public interface Player {
    */
   boolean isCommandAvailable(@Command int command);
 
+  /**
+   * Returns the player's currently available {@link Commands}.
+   *
+   * <p>The returned {@link Commands} are not updated when available commands change. Use {@link
+   * EventListener#onAvailableCommandsChanged(Commands)} to get an update when the available
+   * commands change.
+   *
+   * <p>Executing a command that is not available (for example, calling {@link #next()} if {@link
+   * #COMMAND_SEEK_TO_NEXT_MEDIA_ITEM} is unavailable) will neither throw an exception nor generate
+   * a {@link #getPlayerError()} player error}.
+   *
+   * <p>{@link #COMMAND_SEEK_TO_NEXT_MEDIA_ITEM} and {@link #COMMAND_SEEK_TO_PREVIOUS_MEDIA_ITEM}
+   * are unavailable if there is no such {@link MediaItem}.
+   *
+   * @return The currently available {@link Commands}.
+   * @see EventListener#onAvailableCommandsChanged
+   */
+  Commands getAvailableCommands();
+
   /** Prepares the player. */
   void prepare();