+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayerImpl.java
@@ -340,8 +340,8 @@ import java.util.concurrent.CopyOnWriteArraySet;
   }
 
   @Override
-  public boolean isCommandAvailable(@Command int command) {
-    return availableCommands.contains(command);
+  public Commands getAvailableCommands() {
+    return availableCommands;
   }
 
   @Override